code refactoring

+++ src/blog/blog.controller.ts
@@ -1,13 +1,31 @@
-import { Controller, Post, Body, Req, UseGuards, HttpException, HttpStatus, Param, Patch, ParseIntPipe, BadRequestException, InternalServerErrorException, Get, Query, NotFoundException, ForbiddenException, Delete } from '@nestjs/common';
+import {
+  Controller,
+  Post,
+  Body,
+  Req,
+  UseGuards,
+  Param,
+  Patch,
+  ParseIntPipe,
+  BadRequestException,
+  InternalServerErrorException,
+  Get,
+  Query,
+  NotFoundException,
+  ForbiddenException,
+  Delete,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { BlogService } from './blog.service';
 import { AuthGuard } from '@nestjs/passport';
 import { CreateBlogDto } from './dto/create-blog.dto';
-import { Blog } from './entities/blog.entity';
 import { AuthenticatedRequest } from 'src/types/authenticated-request.interface';
 import { UpdateBlogDto } from './dto/update-blog.dto';
-import { CurrentUser } from '@app/auth/decorators/current-user.decorator';
 import { Public } from '@app/auth/decorators/public.decorator';
 import { UpdateBlogStateDto } from './dto/update-blog-state.dto';
+import { ApiOkResponse } from '@nestjs/swagger';
+import { BlogResponseDto } from './dto/blog-response.dto';
+import { MyBlogResponseDto } from './dto/my-blog-response.dto';
 
 @Controller('blogs')
 export class BlogController {
@@ -17,71 +35,21 @@ export class BlogController {
   @UseGuards(AuthGuard('jwt'))
   async createArticle(
     @Body() createArticleDto: CreateBlogDto,
-    @Req() req: any,
+    @Req() req: AuthenticatedRequest,
   ) {
-    const userId = req.user.id;
-    const savedArticle = await this.blogService.createArticle(
-      createArticleDto,
-      userId,
-    );
-    return { statusCode: 201, data: savedArticle };
-  }
-
-  @Patch(':id/state')
-  @UseGuards(AuthGuard('jwt'))
-  async updateArticleState(
-    @Param('id') id: number,
-    @Body() updateBlogStateDto: UpdateBlogStateDto,
-    @Req() req: AuthenticatedRequest,
-  ): Promise<{ status: boolean; article: Blog }> {
-    const userId = req.user._id;
-
-    try {
-      const updatedArticle = await this.blogService.updateArticleState(
-        id,
-        userId,
-        updateBlogStateDto.state,
-      );
-      return { status: true, article: updatedArticle };
-    } catch (error) {
-      if (error instanceof HttpException) {
-        throw error;
-      }
-      console.error('Error updating article state:', error);
-      throw new HttpException(
-        'Internal Server Error',
-        HttpStatus.INTERNAL_SERVER_ERROR,
-      );
-    }
-  }
-
-  @UseGuards(AuthGuard('jwt'))
-  @Patch(':id')
-  async updateArticle(
-    @Param('id', ParseIntPipe) id: number,
-    @Req() req: AuthenticatedRequest,
-    @Body() updates: UpdateBlogDto,
-  ): Promise<Blog> {
-    const userId = req.user._id;
-
-    const result = await this.blogService.getArticleById(id);
-    const article = result.article;
-
-    if (article?.author?.id !== userId) {
-      throw new ForbiddenException(
-        'You do not have permission to edit this article',
-      );
-    }
-
-    return await this.blogService.editArticle(id, userId, updates);
+    const userId = req.user.userId;
+    return await this.blogService.createArticle(createArticleDto, userId);
   }
 
   @Public()
   @Get()
+  @ApiOkResponse({
+    type: BlogResponseDto,
+    isArray: true,
+  })
   async getArticles(@Query() query) {
     try {
-      const articles = await this.blogService.getArticles(query);
-      return { status: true, articles };
+      return await this.blogService.getArticles(query);
     } catch (error) {
       if (error instanceof BadRequestException) {
         throw new BadRequestException(error.message);
@@ -93,12 +61,76 @@ export class BlogController {
     }
   }
 
+  @UseGuards(AuthGuard('jwt'))
+  @Get('my-articles')
+  async getUserArticles(
+    @Req() req: AuthenticatedRequest,
+  ): Promise<MyBlogResponseDto[]> {
+    const user = req.user;
+
+    if (!user || !user.userId) {
+      throw new UnauthorizedException('User not authenticated');
+    }
+
+    console.log('Authenticated user ID:', user.userId);
+
+    const articles = await this.blogService.getUserArticles(user.userId);
+
+    return articles.length ? articles : [];
+  }
+
+  @Patch(':id/state')
+  @UseGuards(AuthGuard('jwt'))
+  async updateArticleState(
+    @Param('id', ParseIntPipe) id: number,
+    @Body() updateBlogStateDto: UpdateBlogStateDto,
+    @Req() req: AuthenticatedRequest,
+  ): Promise<MyBlogResponseDto> {
+    console.log('req.user:', req.user);
+
+    const userId = req.user.userId;
+    console.log('User ID:', userId);
+
+    return await this.blogService.updateArticleState(
+      id,
+      userId,
+      updateBlogStateDto.state,
+    );
+  }
+
+  @Patch(':id')
+  @UseGuards(AuthGuard('jwt'))
+  async updateArticle(
+    @Param('id', ParseIntPipe) id: number,
+    @Req() req: AuthenticatedRequest,
+    @Body() updates: UpdateBlogDto,
+  ): Promise<MyBlogResponseDto> {
+    const userId = req.user.userId;
+
+    const article = await this.blogService.findOneById(id);
+
+    if (!article) {
+      throw new NotFoundException('Article not found');
+    }
+
+    if (article.author.id !== userId) {
+      throw new ForbiddenException(
+        'You do not have permission to edit this article',
+      );
+    }
+
+    return await this.blogService.editArticle(id, userId, updates);
+  }
+
   @Public()
   @Get(':id')
+  @ApiOkResponse({
+    type: BlogResponseDto,
+    isArray: false,
+  })
   async getArticleById(@Param('id') id: number) {
     try {
-      const result = await this.blogService.getArticleById(id);
-      return result;
+      return await this.blogService.getArticleById(id);
     } catch (error) {
       if (error instanceof NotFoundException) {
         throw error;
@@ -111,17 +143,18 @@ export class BlogController {
 
   @UseGuards(AuthGuard('jwt'))
   @Delete(':id')
-  async deleteArticleById(@Param('id') id: number, @Req() req) {
-    const userId = req.user.id;
+  async deleteArticleById(
+    @Param('id') id: number,
+    @Req() req: AuthenticatedRequest,
+  ) {
+    const userId = req.user?.userId;
+
+    if (!userId) {
+      throw new ForbiddenException('User not authenticated');
+    }
 
     try {
-      const article = await this.blogService.deleteArticleById(id, userId);
-
-      return {
-        status: true,
-        message: 'Article deleted successfully',
-        article,
-      };
+      return await this.blogService.deleteArticleById(id, userId);
     } catch (error) {
       if (error instanceof NotFoundException) {
         throw new NotFoundException(error.message);
@@ -132,19 +165,4 @@ export class BlogController {
       throw error;
     }
   }
-
-  @UseGuards(AuthGuard('jwt'))
-  @Get('my-articles')
-  async getUserArticles(@CurrentUser() user: any) {
-    const userId = user.id;
-
-    const articles = await this.blogService.getUserArticles(userId);
-
-    return {
-      status: true,
-      message: 'Articles fetched successfully',
-      articles,
-    };
-  }
 }
-
